+++ priv/repo/migrations/20241121160003_create_collection_items_sequence.exs
@@ -1,0 +1,23 @@
+defmodule Lightning.Repo.Migrations.CollectionItemsSequence do
+  use Ecto.Migration
+  @disable_migration_lock true
+  @disable_ddl_transaction true
+
+  def up do
+    execute("CREATE SEQUENCE collection_items_id_seq")
+
+    alter table(:collection_items) do
+      modify :id, :bigint,
+        null: false,
+        default: fragment("nextval('collection_items_id_seq'::regclass)")
+    end
+  end
+
+  def down do
+    alter table(:collection_items) do
+      modify :id, :bigint, null: true, default: nil
+    end
+
+    execute("DROP SEQUENCE collection_items_id_seq")
+  end
+end